llvm: readme: update links
A few projects have moved to newer versions of Clang, update the status.
(Observed from the links provided).

Cuttlefish configs are a better source of truth for Android common
kernels.  Action item to upgrade hikey toolchain.

Delete links to old prebuilts that have been removed, as the links are
currently broken.  Alternatively, we could keep the links, but update
the branch they refer to.

Bug: 120503084
Change-Id: Ife44632f3ae1c8da34ba6542756ac5224666716c
Signed-off-by: Nick Desaulniers <ndesaulniers@google.com>

Former-commit-id: 23dbfac5b7e454dc87fa8446f170828d49f434e8